se ha modificado ProductoService por problemas e inconsistencias con Productoservice.Todo funciona bien

+++ src/main/java/com/ecommerce/backendnpu/service/ProductoService.java
@@ -4,13 +4,12 @@ import com.ecommerce.backendnpu.model.Producto;
 import java.util.List;
 import java.util.Optional;
 
-
 public interface ProductoService {
     List<Producto> getAllProductos();
-    Optional<Producto> getProductoById(Integer id);
+    Optional<Producto> getProductoById(Long id);
     Producto saveProducto(Producto producto);
     Producto updateProducto(Producto producto);
-    void deleteProducto(Integer id);
+    void deleteProducto(Long id);
     List<Producto> searchProductos(String nombre);
     List<Producto> filterProductosByCategoria(Integer categoriaId);
 }
